main driver complete

+++ expense-tracker/migrations/20251205173409_create_expense_table.sql
@@ -1,7 +1,7 @@
 -- Add migration script here
 CREATE TABLE IF NOT EXISTS expenses (
     id UUID PRIMARY KEY NOT NULL,
-    amount DECIMAL(10, 2) NOT NULL,
+    amount DOUBLE PRECISION NOT NULL,
     category TEXT NOT NULL,
     description TEXT NOT NULL,
     date DATE NOT NULL
